requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3